Temporarily remove validator from image and use openjdk
The earlier fix to use the new validator/validator base image from
GitHub exposed an issue that their usage of distroless means basic
assumptions like the existence of bash are no longer valid. This is
a quick fix to just stick with the openjdk base for the time being
to unblock a pipeline deployment since validator isn't currently
used.